Remove the operation module + reorganize cli module
This commit is a rehaul to make the cli code easier to follow. The
functionality of some cli features is now removed, and will stay that
way for now. In the future, the functionality will be done with tests,
and once those tests are working, the cli will be connected to it.

It also drops the Operation module, which is redundant with a
database centric workflow. I think this should make it easier to follow
along with what Mix is doing internally.

The documentation has not been updated, this will be done in the future.